Add files via upload
The above code works as follows:

The program asks the user to enter a text and stores that text in the text variable.

Then, it calls the function countWords(text) passing the entered text as an argument.

The countWords(text) function runs through each character of the text. If it finds a letter or digit, it assumes it is within a word. When it encounters a space, punctuation, or any other non-word character, it increments the numberWords counter.

Finally, the function returns the result of the word count.

This is a simple and basic example of a word counter in Java. It is important to note that this code does not take into account some special situations, such as hyphenated words or special characters, and could be improved to deal with more complex cases.